Updated ID token parsing logic

+++ IdentityCore/src/oauth2/token/MSIDAccessToken.m
@@ -31,7 +31,7 @@
 #import "MSIDConfiguration.h"
 
 //in seconds, ensures catching of clock differences between the server and the device
-static uint64_t s_expirationBuffer = 300;
+static NSUInteger s_expirationBuffer = 300;
 
 @interface MSIDAccessToken()
